Hey, I got the images working better, and thumbnails now make page loading very quick. Uploading stuff to the site will still not be the fastest. I have had to host the application on another server though. http://www.randalwa.com.au/archives/
This link is in my magic lamp site under portfolio. There is a list of my work and
Mapping Archeologist
will take you to the archeologist site. You can upload a relic with just a name, then later you can go to the admin interface
http://www.randalwa.com.au/admin/archives/relic/
to add any further information.

For some reason I just uploaded another relic, and the photo thumbnail is not showing. I will look into this, but in the mean time
it should work okay for showing. The application senses the size of the screen you are using, and resizes the map accordingly.
On some screens the map is too big, but I think I have fixed that now. Email me if it is still aproblem.
Cheers
Jason